requirements_dev.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
cloudinary==1.28.0
cssselect==1.1.0
Django==3.2.9
djangorestframework==3.13.1
django-ckeditor==6.2.0
django-cookie-consent==0.3.0
django-crispy-forms==1.14.0
django-cors-headers==3.11.0
django-js-asset==1.2.2
django-taggit==2.0.0
django-jazzmin==2.4.9
django-hashers-passlib==0.4
lxml==4.6.5
markdown2==2.4.2
passlib==1.7.4
Pillow==8.4.0
psycopg2_binary==2.9.2
pyquery==1.4.3
python-decouple==3.5
pytz==2021.3
readtime==1.1.1
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2